lvmd: Move test utils to sep. package
Add function to create a LV in a VG too.

So we can leverage in a couple of different sub packages.
